prefetch gif before showing
